Remove 'Member' interface
