+++ src/main/java/minijava/ast/Class.java
@@ -1,14 +1,24 @@
 package minijava.ast;
 
+import java.util.Collections;
 import java.util.List;
 
 public class Class<TRef> {
   public final String name;
-  public final List<Member<TRef>> members;
+  public final List<Field<TRef>> fields;
+  public final List<Method<TRef>> methods;
 
-  public Class(String name, List<Member<TRef>> members) {
+  /**
+   * Constructs a new class node.
+   *
+   * <p><strong>We do <em>not</em> make defensive copies</strong> of {@code fields} or {@code
+   * methods}. The caller must make sure that, after handing over these lists, no modifications
+   * happen to them.
+   */
+  public Class(String name, List<Field<TRef>> fields, List<Method<TRef>> methods) {
     this.name = name;
-    this.members = members;
+    this.fields = Collections.unmodifiableList(fields);
+    this.methods = Collections.unmodifiableList(methods);
   }
 
   public <TRet> TRet acceptVisitor(Visitor<TRef, TRet> visitor) {
@@ -16,7 +26,6 @@ public class Class<TRef> {
   }
 
   public interface Visitor<TRef, TReturn> {
-
     TReturn visitClassDeclaration(Class<TRef> that);
   }
 }
